refactor: extract IDERunManager from SikulixIDE (Phase 3, Step 2)
Move ButtonRun, ButtonRunViz inner classes and run-related methods
(addErrorMark, resetErrorMark, installStopHotkey, onStopRunning) to
IDERunManager.java. Cleaned 3 unused imports from SikulixIDE.

SikulixIDE.java: 2344 -> 2196 lines (-148)
IDERunManager.java: 228 lines (new)
